created account row component and fixed pr comments
