+++ setup/react/app/variables.css
@@ -181,6 +181,7 @@ or "warn/action" ineastd of "red/green"
   --color-tx-priority-btn-default: var(--color-maastricht-blue);
   --color-progress-bar: var(--color-ink-blue);
   --color-ghost-white: #f7f9fb;
+  --color-button-outline-color: var(--color-very-light-gray);
 
   /*************************
         Box
@@ -270,4 +271,5 @@ or "warn/action" ineastd of "red/green"
   --color-progress-bar: var(--color-jade-green);
   --color-zodiac-blue: #0c152e;
   --color-background-dropdown: var(--color-body-bg);
+  --color-button-outline-color: #4c4c4c;
 }
