+++ src/modules/account/components/AccountRow/index.js
@@ -1,0 +1,1 @@
+export { default } from './AccountRow';
