+++ src/modules/account/components/AccountSelect/AccountSelect.css
@@ -53,48 +53,6 @@
   margin-top: 16px;
 }
 
-.accountWraper:last-child {
-  border-bottom: none;
-}
-
-.accountWraper:hover {
-  opacity: 0.8;
-}
-
-.accountWraper {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 0px;
-  padding-bottom: 16px;
-  border: none;
-  border-radius: 0px;
-  justify-content: flex-start;
-  color: var(--color-maastricht-blue);
-  border-bottom: 1px solid#dfe6f2;
-  transition: opacity 100ms linear;
-  cursor: pointer;
-
-  & > div:first-child ~ div {
-    margin-left: 10px;
-    flex: 1;
-  }
-
-  & > div:first-child ~ div > p {
-    margin-top: 9px;
-    margin-bottom: 0px;
-    color: var(--color-slate-gray);
-    font-size: 13px;
-  }
-
-  & > div:first-child ~ div ~ button {
-    border: none;
-    background-color: transparent;
-    width: 40px;
-  }
-}
-
 .accountSelect {
   align-items: center;
   display: flex;
