+++ src/modules/account/components/AccountSelect/AccountSelect.js
@@ -4,48 +4,9 @@ import Box from 'src/theme/box';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import { OutlineButton } from 'src/theme/buttons';
 import Icon from 'src/theme/Icon';
-import WalletVisual from '@wallet/components/walletVisual';
 import { useAccounts } from '../../hooks/useAccounts';
 import styles from './AccountSelect.css';
-
-function AccountRow({
-  account,
-  onSelect,
-  onRemove,
-  showRemove,
-}) {
-  const { uuid, metadata: { name, address } } = account;
-
-  return (
-    <div
-      key={uuid}
-      data-testid={uuid}
-      className={styles.accountWraper}
-      onClick={() => onSelect(account)}
-    >
-      <WalletVisual address={address} size={40} />
-      <div align="left">
-        <b className={`${styles.addressValue}`}>
-          {name}
-        </b>
-        <p className={`${styles.addressValue}`}>
-          {address}
-        </p>
-      </div>
-      {showRemove && (
-        <button
-          data-testid="delete-icon"
-          onClick={(event) => {
-            event.stopPropagation();
-            onRemove(account);
-          }}
-        >
-          <Icon name="deleteRedIcon" />
-        </button>
-      )}
-    </div>
-  );
-}
+import AccountRow from '../AccountRow';
 
 const AccountSelect = ({ onSelectAccount, onAddAccount, onRemoveAccount }) => {
   const { t } = useTranslation();
